doble click - abrir panel y cancelEdit

+++ services/aprendoz_test/src/com/aprendoz_test/data/PersonaEdadId.java
@@ -6,7 +6,7 @@ import java.io.Serializable;
 
 /**
  *  aprendoz_test.PersonaEdadId
- *  05/09/2014 07:43:03
+ *  07/15/2014 14:36:07
  * 
  */
 public class PersonaEdadId
@@ -14,12 +14,12 @@ public class PersonaEdadId
 {
 
     private Integer idPersona;
-    private Long edad_vistaAlumnosActivos_idPersona_;
+    private Integer edad_vistaAlumnosActivos_idPersona_;
 
     public PersonaEdadId() {
     }
 
-    public PersonaEdadId(Integer idPersona, Long edad_vistaAlumnosActivos_idPersona_) {
+    public PersonaEdadId(Integer idPersona, Integer edad_vistaAlumnosActivos_idPersona_) {
         this.idPersona = idPersona;
         this.edad_vistaAlumnosActivos_idPersona_ = edad_vistaAlumnosActivos_idPersona_;
     }
@@ -74,11 +74,11 @@ public class PersonaEdadId
         this.idPersona = idPersona;
     }
 
-    public Long getEdad_vistaAlumnosActivos_idPersona_() {
+    public Integer getEdad_vistaAlumnosActivos_idPersona_() {
         return edad_vistaAlumnosActivos_idPersona_;
     }
 
-    public void setEdad_vistaAlumnosActivos_idPersona_(Long edad_vistaAlumnosActivos_idPersona_) {
+    public void setEdad_vistaAlumnosActivos_idPersona_(Integer edad_vistaAlumnosActivos_idPersona_) {
         this.edad_vistaAlumnosActivos_idPersona_ = edad_vistaAlumnosActivos_idPersona_;
     }
